# api/docs/swagger.yaml
definitions:
...
  entity.CreateMerchantBranchRequest:
    properties:
      address:
        type: string
      merchant_id:
        type: string
      name:
        type: string
    type: object
  entity.CreateMerchantBranchResponse:
    properties:
      guid:
        type: string
    type: object
  entity.CreateMerchantRequest:
    properties:
      background_image:
        type: string
      comission:
        type: number
      delivery_time:
        type: number
      description:
        type: string
      logo:
        type: string
      name:
        type: string
      status:
        type: boolean
    type: object
  entity.CreateMerchantResponse:
    properties:
      guid:
        type: string
    type: object
...
  entity.DeleteMerchantBranchRequest:
    properties:
      guid:
        type: string
    type: object
  entity.DeleteMerchantRequest:
    properties:
      guid:
        type: string
    type: object
...
  entity.GetMerchantBranchOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/entity.GetOrderResponse'
        type: array
    type: object
  entity.GetMerchantBranchResponse:
    properties:
      address:
        type: string
      created_at:
        type: string
      guid:
        type: string
      merchant_id:
        type: string
      name:
        type: string
    type: object
  entity.GetMerchantBranchesResponse:
    properties:
      merchant_branches:
        items:
          $ref: '#/definitions/entity.GetMerchantBranchResponse'
        type: array
    type: object
  entity.GetMerchantOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/entity.GetOrderResponse'
        type: array
    type: object
  entity.GetMerchantProductsResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/entity.GetProductResponse'
        type: array
    type: object
  entity.GetMerchantResponse:
    properties:
      background_image:
        type: string
      comission:
        type: number
      created_at:
        type: string
      delivery_time:
        type: number
      description:
        type: string
      guid:
        type: string
      logo:
        type: string
      name:
        type: string
      status:
        type: boolean
    type: object
  entity.GetOrderResponse:
    properties:
      address:
        type: string
      branch_address:
        type: string
      branch_id:
        type: string
      branch_name:
        type: string
      comission:
        type: number
      comment:
        type: string
      courier_id:
        type: string
      created_at:
        type: string
      delivery_price:
        type: number
      delivery_time:
        type: number
      delivery_type:
        type: string
      guid:
        type: string
      image:
        type: string
      merchant_name:
        type: string
      payment_type:
        type: string
      products:
        items:
          $ref: '#/definitions/entity.OrderProducts'
        type: array
      status:
        type: string
      user_id:
        type: string
    type: object
  entity.GetProductResponse:
    properties:
      category_id:
        type: string
      created_at:
        type: string
      guid:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.Option'
        type: array
      price:
        type: number
    type: object
...
  entity.ListMerchantBranchesResponse:
    properties:
      merchant_branches:
        items:
          $ref: '#/definitions/entity.GetMerchantBranchResponse'
        type: array
    type: object
  entity.ListMerchantResponse:
    properties:
      merchants:
        items:
          $ref: '#/definitions/entity.GetMerchantResponse'
        type: array
    type: object
...
  entity.Option:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
...
  entity.OrderProducts:
    properties:
      category_id:
        type: string
      count:
        type: integer
      created_at:
        type: string
      guid:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.Option'
        type: array
      price:
        type: number
    type: object
  entity.ResponseOK:
    properties:
      message:
        type: object
    type: object
...
  entity.UpdateMerchantBranchRequest:
    properties:
      address:
        type: string
      guid:
        type: string
      merchant_id:
        type: string
      name:
        type: string
    type: object
  entity.UpdateMerchantBranchResponse:
    properties:
      guid:
        type: string
    type: object
  entity.UpdateMerchantRequest:
    properties:
      background_image:
        type: string
      comission:
        type: number
      delivery_time:
        type: number
      description:
        type: string
      guid:
        type: string
      logo:
        type: string
      name:
        type: string
      status:
        type: boolean
    type: object
  entity.UpdateMerchantResponse:
    properties:
      guid:
        type: string
    type: object
...
  models.ResponseError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/merchant-branch/{guid}/orders:
    get:
      consumes:
      - application/json
      description: API that get merchant-branch orders
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetMerchantBranchOrdersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get merchant-branch orders
      tags:
      - merchant-branch
  /v1/merchant-branch/create:
    post:
      consumes:
      - application/json
      description: API that create merchant-branch
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateMerchantBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CreateMerchantBranchResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create merchant-branch
      tags:
      - merchant-branch
  /v1/merchant-branch/delete:
    delete:
      consumes:
      - application/json
      description: API that delete merchant-branch
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteMerchantBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseOK'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete merchant-branch
      tags:
      - merchant-branch
  /v1/merchant-branch/get/{guid}:
    get:
      consumes:
      - application/json
      description: API that get merchant-branch
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetMerchantBranchResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get merchant-branch
      tags:
      - merchant-branch
  /v1/merchant-branch/list:
    get:
      consumes:
      - application/json
      description: API list merchants-branches
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ListMerchantBranchesResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List merchant-branches
      tags:
      - merchant-branch
  /v1/merchant-branch/update:
    put:
      consumes:
      - application/json
      description: API that update merchant-branch
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateMerchantBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdateMerchantBranchResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update merchant-branch
      tags:
      - merchant-branch
  /v1/merchant/create:
    post:
      consumes:
      - application/json
      description: API that create merchant
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateMerchantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CreateMerchantResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create merchant
      tags:
      - merchant
  /v1/merchant/delete:
    delete:
      consumes:
      - application/json
      description: API that delete merchant
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteMerchantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseOK'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete merchant
      tags:
      - merchant
  /v1/merchant/get-branches/{guid}:
    get:
      consumes:
      - application/json
      description: API that get merchant-branches
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetMerchantBranchesResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get merchant-branches by merchant id
      tags:
      - merchant
  /v1/merchant/get/{guid}:
    get:
      consumes:
      - application/json
      description: API that get merchant
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetMerchantResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get merchant
      tags:
      - merchant
  /v1/merchant/list:
    get:
      consumes:
      - application/json
      description: API list merchants
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ListMerchantResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List merchants
      tags:
      - merchant
  /v1/merchant/orders/{guid}:
    get:
      consumes:
      - application/json
      description: API that get merchant-orders
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: branch_name
        in: query
        name: branch_name
        type: string
      - description: user_name
        in: query
        name: user_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetMerchantOrdersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get merchant-orders by merchant id
      tags:
      - merchant
  /v1/merchant/products/{guid}:
    get:
      consumes:
      - application/json
      description: API that get merchant-products
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetMerchantProductsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get merchant-products by merchant id
      tags:
      - merchant
  /v1/merchant/update:
    put:
      consumes:
      - application/json
      description: API that update merchant
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateMerchantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdateMerchantResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update merchant
      tags:
      - merchant
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"